+++ scripts/mdbook-forc-documenter/src/plugins.rs
@@ -1,0 +1,3 @@
+pub fn official_plugin_commands() -> Vec<String> {
+    vec!["fmt".to_string(), "explore".to_string(), "lsp".to_string()]
+}
